+++ lib/shoryuken/runner.rb
@@ -9,10 +9,6 @@ require 'erb'
 require 'shoryuken'
 
 module Shoryuken
-  # Exception raised to trigger shutdown
-  # @see https://github.com/mperham/sidekiq/blob/33f5d6b2b6c0dfaab11e5d39688cab7ebadc83ae/lib/sidekiq/cli.rb#L20
-  class Shutdown < Interrupt; end
-
   # Runs the Shoryuken server process.
   # Handles signal trapping, daemonization, and lifecycle management.
   class Runner